zendesk#3204: Fix "lastseenexpireafter" 32-bit signed int overflow.
Changelog: title
(cherry picked from commit d68441c51641cbcaeb12c117d31a15ee2dcc6c74)

+++ libpromises/cf3.extern.h
@@ -69,7 +69,7 @@ extern const char *const CF_AGENTTYPES[];
 extern int CFA_MAXTHREADS;
 extern AgentType THIS_AGENT_TYPE;
 extern int SHOWREPORTS;
-extern int LASTSEENEXPIREAFTER;
+extern long LASTSEENEXPIREAFTER;
 extern const char *DEFAULT_COPYTYPE;
 extern Item *PROCESSTABLE;
